commit2 pointtext

+++ pointtext.go
@@ -1,10 +1,9 @@
 package main
 
 
-
 type Point struct {
-	X    int
-	Y    int
+	X    float64
+	Y    float64
 	Text string
 }
 
@@ -12,6 +11,6 @@ func PointText(p Point) Point {
 	return Point{
 		X:    p.X,
 		Y:    p.Y,
-		Text: fmt.Sprintf("Text at (%d, %d)", p.X, p.Y),
+		Text: fmt.Sprintf("Text at (%g, %g)", p.X, p.Y),
 	}
 }
